ESQL: Merge Order implemenations
This merges esql-core's `Order` into esql's `Order`. We don't need both.
